refactor: move dataset hook tests out of ui
