+++ crates/fricon-ui/frontend/src/features/datasets/hooks/useDatasetTableSelection.ts
